refactor: remove auth.js and favour concentrating code in http-client.js
