+++ actions/http-client.js
@@ -11,7 +11,65 @@ governing permissions and limitations under the License.
 */
 
 const got = require("got");
-const { getAuthProviderFromParams } = require("./auth");
+const { Core } = require("@adobe/aio-sdk");
+const authLogger = Core.Logger("auth", { level: "info" });
+const {
+  imsProviderWithEnvResolver,
+  integrationProviderWithEnvResolver,
+} = require("../utils/adobe-auth");
+const {
+  CommerceSdkValidationError,
+} = require("@adobe/aio-commerce-lib-core/error");
+
+/**
+ * This function returns the auth function from @adobe/aio-commerce-lib-auth based on the environment parameters.
+ * @param {object} params - Environment params from the IO Runtime request
+ * @returns the auth object for the request
+ * @throws {Error} - throws error if the params are missing
+ */
+async function getAuthProviderFromParams(params) {
+  // `aio app dev` compatibility: inputs mapped to undefined env vars come as $<input_name> in dev mode, but as '' in prod mode
+  try {
+    if (
+      params.COMMERCE_CONSUMER_KEY &&
+      params.COMMERCE_CONSUMER_KEY !== "$COMMERCE_CONSUMER_KEY"
+    ) {
+      authLogger.info("Commerce client will use CommerceIntegration provider");
+      const integrationProvider =
+        await integrationProviderWithEnvResolver(params);
+      return ({ method, url }) => {
+        return integrationProvider.getHeaders(method, url);
+      };
+    }
+
+    // `aio app dev` compatibility: inputs mapped to undefined env vars come as $<input_name> in dev mode, but as '' in prod mode
+    if (
+      params.OAUTH_CLIENT_ID &&
+      params.OAUTH_CLIENT_ID !== "$OAUTH_CLIENT_ID"
+    ) {
+      authLogger.info("Commerce client will use ImsAuth provider");
+      const imsProvider = await imsProviderWithEnvResolver(params);
+      return async () => {
+        const token = await imsProvider.getAccessToken();
+        return { Authorization: `Bearer ${token}` };
+      };
+    }
+  } catch (error) {
+    if (error instanceof CommerceSdkValidationError) {
+      authLogger.error(
+        `Unable to create authProvider params: ${error.display(false)}`,
+      );
+
+      throw new Error(
+        `Unable to create authProvider params: ${error.display(false)}`,
+      );
+    }
+  }
+
+  throw new Error(
+    "Unknown auth type, supported IMS OAuth or Commerce OAuth1. Please review documented auth types",
+  );
+}
 
 /**
  * This function return the Adobe commerce OAuth client
@@ -138,4 +196,5 @@ async function getClient(clientOptions, logger) {
 
 module.exports = {
   getClient,
+  getAuthProviderFromParams,
 };
